+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project are documented here. This project adheres to
 
 ## Unreleased
 
+## [0.52.169] - 2026-09-01
+
 ### MCP
 
 #### Fixed